extract calender export to a component
